+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/service/RecordQosServiceImpl.java
@@ -153,6 +153,7 @@ public class RecordQosServiceImpl implements RecordQosService {
     private ConsumerStatus internalGetConsumerStatus(ConsumerId consumerId) throws HandlerException {
         ConsumeHandler<? extends Bean> consumeHandler = consumeHandlerMap.get(consumerId);
         return new ConsumerStatus(
+                consumeHandler.bufferedSize(),
                 consumeHandler.getBufferSize(),
                 consumeHandler.getBatchSize(),
                 consumeHandler.getMaxIdleTime(),